requirement.txt
alipay==0.7.4
certifi==2019.9.11
chardet==3.0.4
Django==1.11
django-redis==4.10.0
idna==2.8
mysqlclient==1.4.4
Naked==0.1.31
pycryptodome==3.9.0
pytz==2019.3
PyYAML==5.1.2
redis==3.3.5
requests==2.22.0
shellescape==3.4.1
six==1.12.0
urllib3==1.25.6
